+++ packages/core/src/index.js
@@ -41,6 +41,7 @@ export { default as HvVerticalNavigation } from "./VerticalNavigation";
 export { default as HvAssetInventory } from "./AssetInventory";
 export { default as HvDatePicker } from "./DatePicker";
 export { default as HvFileUploader } from "./FileUploader";
+export { default as HvEmptyState } from "./EmptyState";
 
 // theme
 export { default as hvTheme } from "./theme";