+++ packages/doc/samples/components/checkbox/checkboxState.js
@@ -1,50 +1,53 @@
-import React from "react";
-import HvCheckBox from "@hv/uikit-react-core/dist/Selectors/CheckBox";
+import React, { useState } from "react";
+import { HvCheckBox } from "@hv/uikit-react-core/dist/Selectors";
 
-class CheckboxStateManager extends React.Component {
-  state = {
-    isAChecked: false,
-    isBChecked: false,
-    isCChecked: false
-  };
+const CheckboxStateManager = () => {
+  const [aChecked, setAChecked] = useState(false);
+  const [bChecked, setBChecked] = useState(true);
+  const [cChecked, setCChecked] = useState(false);
 
-  render() {
-    const { isAChecked, isBChecked, isCChecked } = this.state;
+  return (
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "column",
+        alignItems: "flex-start"
+      }}
+    >
+      <HvCheckBox
+        id="checkState1"
+        label="Checkbox 1"
+        checked={aChecked}
+        onChange={() => setAChecked(!aChecked)}
+      />
+      <HvCheckBox
+        id="checkState2"
+        label="Checkbox 2"
+        checked={bChecked}
+        onChange={() => setBChecked(!bChecked)}
+      />
+      <HvCheckBox
+        id="checkState3"
+        label="Checkbox indeterminate"
+        indeterminate={cChecked}
+        onChange={() => setCChecked(!cChecked)}
+      />
 
-    return (
-      <>
-        <div>
-          <HvCheckBox
-            label="some label"
-            checked={isAChecked}
-            onChange={() =>
-              this.setState({ isAChecked: !this.state.isAChecked })
-            }
-          />
-        </div>
-
-        <div>
-          <HvCheckBox
-            label="other label"
-            checked={isBChecked}
-            onChange={() =>
-              this.setState({ isBChecked: !this.state.isBChecked })
-            }
-          />
-        </div>
-
-        <div>
-          <HvCheckBox
-            label="indeterminate"
-            indeterminate={isCChecked}
-            onChange={() =>
-              this.setState({ isCChecked: !this.state.isCChecked })
-            }
-          />
-        </div>
-      </>
-    );
-  }
-}
+      <HvCheckBox
+        id="checkState4"
+        label="Checkbox indeterminate disabled"
+        disabled
+        indeterminate
+      />
+      <HvCheckBox id="checkState5" label="Checkbox disabled" disabled />
+      <HvCheckBox
+        id="checkState6"
+        label="Checkbox checked disabled"
+        checked
+        disabled
+      />
+    </div>
+  );
+};
 
 export default <CheckboxStateManager />;