+++ packages/doc/stories/3-components/kpi.js
@@ -53,6 +53,21 @@ storiesOf("Components/Visualization", module).add("KPI", () => <HvKpi />, {
       title: "6. Nodes",
       description: "A kpi of the number nodes",
       src: "components/kpi/kpiNodes"
+    },
+    {
+      title: "7. Total number of events with icon",
+      description: "A selectable kpi with the total number of events",
+      src: "components/kpi/kpiTNE"
+    },
+    {
+      title: "8. Total number of events",
+      description: "A selectable kpi with the total number of events",
+      src: "components/kpi/kpiTNENI"
+    },
+    {
+      title: "9. Caution of events",
+      description: "A selectable kpi with the total number of caution events",
+      src: "components/kpi/kpiCautionEvents"
     }
   ]
 });